[backend] refactor: move connectivity check to hub domain
